added support for parameters when using Query annotation
